+++ src/kanban_tui/views/settings_tab_view.py
@@ -70,6 +70,7 @@ class SettingsView(Vertical):
     @on(Input.Changed)
     @on(Switch.Changed)
     @on(Button.Pressed)
+    @on(Select.Changed)
     def config_changes(self):
         self.config_has_changed = True